backend/requirements.txt
# Backend dependencies for RecruitAI
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Vector DB
qdrant-client==1.7.0
numpy==1.26.3

# Document Processing (simple, reliable)
PyMuPDF==1.23.8
python-docx==1.1.0

# Embeddings
sentence-transformers==2.2.2

# Validation
email-validator==2.1.0
